requirements.txt
# Microsoft Agent Framework - Core and Azure AI Integration
agent-framework>=0.1.0
agent-framework-azure-ai>=0.1.0

# Azure Authentication
azure-identity>=1.15.0

# Testing Framework
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.29.0
nbformat>=5.9.0

# Code Quality and Linting
black>=24.0.0
ruff>=0.3.0

# Environment Management
python-dotenv>=1.0.0
